drbd: Treat intentionally diskless peers as such even they have a bitmap in MD
If a peer that is marked as intentionally diskless AKA "client" but it has
a bitmap allocated in the local meta-data then teat it as peer we do
not have a bitmap for. That matters e.g. when to create new current-uuid
etc.
We leave the bitmap allocated in the MD until it is explicitly de-allocated
with forget-peer.