compat: scripts to run checks on tests
unused-tests.sh looks at all tests and determines if their result
(the #define) is ever used in the code. If it is not, it reports the test
as unused and it can probably be removed.

constant-tests.sh checks the cocci_cache and determines if a test has the
same result on all supported kernels. If it is always set or always unset,
it may be a potential candidate for removal.
However, this could also mean other things, such as that the feature being
tested for is "too new", so that it doesn't show up in any distro kernels
yet.