drbd: Do not re-use drbd_request for sending peer ack
This created confusion with the RCU transfer log. An RCU list traversal
might still have a reference to a request after the kref count has
dropped to zero. It may then get and put the kref again, causing the
request to be "destroyed" again. To prevent this situation, we need to
avoid re-using the kref of the request. Then we can ignore requests
found in a transfer log traversal with kref count zero.