drbd: Add data-generation-UUID modification upon start of resnyc
When a resync operation starts, the resync-target node starts to record
changes caused by the resync against all other diskfull nodes it is
currently not connected to.
It will update its current-uuid at the end of the resync.

What was missing so far, that the bits it sets in the bitmap are not
maked with a base UUID in the bitmap slots.