drbd: Fix resync after a promote on a resync target to a node that joins later
Two aspects need fixing:
First, when on the resync-target a new current-uuid is generated, it is
short-feed-back into peer_device->current_uuid. That is correct but should
only be done for the peers that we have a connection to. It is only
necesary for the peers we resync from, so only do it for them.

Second, the special case for not sending a bitmap_uuid should be limited
to the case that is descibed in the comment. I.e. for resync-targets
only.