drbd: Update my own bitmap-UUIDs towards peers when getting a new current-UUID
When I get a new current UUID in and we know which peers are connected
to the primary, I need to update the bitmap-UUIDs towards my peers.
Those who are connected to the primary, get the same bitmap-UUID
as my current primary, since at this point they are at the same
data version as I am.

This behaviour avoids false detections of split brains scenarios.
An examle:
Nodes: A,B,C where A is diskless. Let B and C disconnect. Create
a new current-uuid on A (become primary). Completely isolate B.
Create a new current-uuid on A (become primary). Now connect B
and C and see that no split-brain gets detected.