compat: coccinelle patches for REQ_* flags
Previously, the REQ_OP_*/REQ_*/BIO_* flags have all been mapped against
each other in a completely unmaintainble ifdef-maze. This changed when
coccinelle patches were added, but this resulted in an unmaintainable
and broken mix of coccinelle patches and parts of the same old ifdef-maze.

To try and fix this once and for all, introduce coccinelle patches to
deal with everything related to REQ_* flags.

This can probably still be cleaned up, and that will happen over the next
few days/weeks.