drbd: Do not send a newly created UUID when writing to the meta-data failed
So far failures in writing to the meta-data results in an immediate detch,
but the code in new_current_uuid() lacked a check on the disk state
between writing the new current UUID to the meta-data and sendit to
the peers.

I think evaluating the return value of drbd_md_sync() is a lot
easier to understand than to have a check on the disk state
in this place.