docker: allow installing to file system
While I'm not a huge fan of this, we can easily support it for the case
the module is built from source. This was requested and is the case for
piraeus. Having this for the "fetch a pkg from LINBIT" case, would
require more work. Wait till somebody requests it. For RHEL this looks
like this:

docker run -it --rm --privileged \
   -e "LB_INSTALL=yes" \
   -v /lib/modules:/lib/modules -v /usr/src:/usr/src:ro \
   rck:rmme