drbd: Fix recent regression; caused both sides to become sync-source
Commit 2ff203a879 did some code deduplication.
Unfortunately the code was inconsistent and with the code-deduplication
the wrong edition was chosen.
The incomplete code before just caused an incomplete list of flags to
get logged. After the code de-duplication it broke a certain case
of the resync handshake. Add back missing cluase.

The regression caused that a resync negotiation ended in both
sides wanting to become sync-source.

It was introduced after 9.0.19, so the regression was never
released.