compat: refactor coccinelle patches
Since the whole REQ_ flag mechanism got to be quite complex in the end, a lot
of older coccinelle rules are now no longer needed, or the new rules are
redundant. Remove some of these, of course being careful not to introduce
regressions in the patch output.

This is part of a larger effort to drastically reduce the size and complexity of
the coccinelle patches.