drbd: Introduce a new option to rr-conflict
rr-conflict got a new option ABS_RETRY_CONNECT or "retry-connect".

That is usefull in a scenario where you use quorum. Consider that you isolate
the node where your service is running and pacemaker moves the service to
partition that keeps quorum.
Now, in case the newtwork outage is shorter than the time it takes to
stop the service on the node that lost quorum, it is usefull if DRBD
keeps retrying to connect.