docker: add missing deps for RHEL8
Looks like some pkgs that are in the RHEL7 "base" are no longer in the
RHEL8 one. Explicitly install kmod (for modprobe/insmod) and dfn-utils
(for yumdownloader).

Also add these explicitly to the RHEL7 Dockerfile. This could make it
easier if we switch to a more minimal image in the future.