drbd: Allow resync of a re-created peer from day0 UUID
In a world of thinly allocated block devices, a peer might be re-created
and reseted to the day0 UUID. (LINSTOR is a candidate for such clever
moves). If we detect that we need to copy in the bitmap from one of
the unallcoated bitmap slots.

Since this patch needs to change how UUIDs are compared and which
UUIDs get sent this functionality needs to be coupled to a new
protocol version (116).

Necessary changes:
* Send the day0 UUID in the uuids110 packet by sending the bitmap_uuid
  of an unallocated peer slot.
* When comparing UUIDs also look into the unallocated peer slots
* When copying the bitmap from one peer_slot to an other also
  allow the source peer_slot to be unallocated, find a unallocated
  bitmap index in that case.