drbd: fix bm_copy_copy_slot() to copy whole bitmap on 64bit archs
So far this had a bug that caused to copy only the first half
of the bitmap on 64 bit architectures. Since it was not used
very seldom this stayed undetected for a long time. With the
next patch it gets used more often and got a coverage in the
drbd9-tests testsuite.