drbd: return to D_UP_TO_DATE upon connection loss while receiving the bitmap
Return to D_UP_TO_DATE if a node was D_OUTDATED due to becoming resync
target. Between the decission to become resync target and receiving the
resync blocks is the bitmap exchange. The disk state is D_OUTDATED during
that time. (Later it advances to D_INCONSISTENT while receiving the
resync blocks).

If we loose connection while doing the bitmap exchange, the disk
state should return to D_UP_TO_DATE if valid.