drbd: Do not set entire bitmap when attaching metadata on DAX/PMEM
When attaching metadata on a block device, we do the following:

1. Resize the in-memory bitmap from zero to the appropriate size for the
   disk being attached.
2. Set the newly assigned memory to all ones.
3. Read the contents from the on-disk bitmap into the in-memory bitmap.

With metadata on DAX/PMEM there is no distinction between the in-memory
and on-disk bitmap. Hence step 2 causes the bitmap to be overwritten.

Fix this bug by skipping step 2 when we are attaching to metadata on a
DAX/PMEM device.