drbd: Do a full peer check before wait_up_to_date() before promote to PRIMARY
Consider a 3 (or more) node setup with quorum enabled. The primary node gets
isolated. It happens that Pacemaker/Corosync detects the failure quicker than
DRBD. It tries to promote one of the remaining nodes.

The surviving nodes will go through the Consistent -> UpToDate mechanism,
they need to find out if there is still a multi-hop connection to the primary.
The surviving nodes will do a reconciliation resync.

The promote should be delayed after these actions.