drbd: Do not assume that a L_PAUSED_SYNC_S has not a D_INCONSISTENT disk
That was right at the time the code was written (drbd-8.4 = only one peer),
but now we can be sync target = D_INCONSISTENT and paused sync source
towards an other peer at the same time.

This bug was really hard to understand since the modification to the
requested state change was not logged.