drbd: Prevent NULL pointer dereference in set_cache_ptr_if_null
This can occur when another thread runs advance_cache_ptr while
set_cache_ptr_if_null is running. Between the calls to cmpxchg in the
thread running set_cache_ptr_if_null, the other thread may advance the
cache pointer to the end of the list and set it to NULL. Hence the
second cmpxchg returns NULL which is then dereferenced to read
dagtag_sector to check which request is older.

The resulting crash was as follows:

  BUG: unable to handle kernel NULL pointer dereference at 0000000000000050
  ...
  RIP: 0010:set_cache_ptr_if_null+0x23/0x40 [drbd]
  ...
  Call Trace:
   mod_rq_state+0x57c/0x6d0 [drbd]
   process_one_request+0x305/0x450 [drbd]