compat: do not apply safety check in rdma module
There was a safety check in place to ensure that all code
transformations of a "set" of changes are applied, because DRBD would
break at runtime otherwise.
This obviously triggers with the rdma code base, since none of the
rules apply.
Therefore, just add a hack-ish "rdma detection" and exclude it from
the check.