compat: amend hint about creating a tgz with SPAAS patch
This hint is really useful for creating a new tarball out of a newly
generated SPAAS patch, but unfortunately it is a bit misleading.
When SPAAS is run, the source tree is obviously patched, so the user would
end up with a patched tree AND the new patch in the tarball, which would
lead to a "Reversed (or previously applied) patch detected!" and subsequently
fail the build.
Thus, add a "make unpatch" to the hint.