compat: fix 8.4 to 9 handshake regression
Regression introduced with DRBD 9.0.19 (and .20):
7d15e1896 drbd: fix wrong state change when current UUID changes during UUID exchange
2ff203a87 drbd: Fix wrong state change when UUID flags change during UUID exchange

DRBD 9 now always compares the "communicated" current uuid and flags
with the current uuid, to double check a change in local UUIDs during
handshake, and if changed, triggers a re-connect.

Only that the "communicated" uuid was never initialized in the 8.4
compat code path, so will always differ :-/

fix: initialize it in the 8.4 compat code path,
and make sure to mask out any "too recent" uuid-flags.