drbd: Correct bitmap UUID after resync
Fixup for 485d3dd6bb30d66a6461d75268afff14e04a0638. This commit fixed a
situation where UUIDs changed during a resync causing the bitmap UUIDs
for other peers to be set wrongly. Unfortunately, it used the current
UUID from the sync source from the start of the resync to populate the
bitmap UUIDs at the end of the resync. In the normal case, when the
current UUID did not change during the resync, this led to a situation
where the bitmap UUID was the same as the current UUID, which is
nonsense.

Fix this by storing our own current UUID at the start of the resync
instead of that from the sync source. This is then used to update the
bitmap UUIDs at the end of the resync.