drbd: Fix pushing bitmap UUID into history when changed
91e95a7cb5c82176b0c8b5f484a675051b033705 introduced a regression that
caused an unexpected split-brain disconnection. The issue was that the
bitmap UUID was updated after pushing the old one into the history.
Since the mentioned commit, UUIDs are not added to the history if they
are in a bitmap UUID slot. As a result, the old bitmap UUID was not
added to the history.

To reproduce the issue with 3 nodes, first arrange for data generations:
* Newest data on A
* Older data on B
* Oldest data on C (but not the initial data generation)
Then:
* Connect A to B
* Connect B to C

C should successfully sync from B. The issue causes a split-brain
disconnection.