drbd: Fix current UUID showing up in bitmap slots unexpectedly
See commit ae3e2c64
drbd: Update my own bitmap-UUIDs towards peers when getting a new current-UUID

It fails to exclude diskless primaries from the described logic. As the
disk of the diskless primary stays unmodified, the logic of draging along
the base UUID of that bitmap slot is wrong. Excluding it is the right thing.

The actual problem is the __set_bitmap_slots() in
drbd_uuid_received_new_current() . The minimal change would be to
remove this diskless primary from the upd mask there.

I think it is right to change the meaning of the weak_nodes to also
consider mark a diskless primary as weak, since it's (detached disk)
is not even weakly connected to the current data.