drbd: Fix logging when usermode helper fails
An unfortunate combination of the ternary operator and macro expansion
caused the log message to be lost whenever the usermode helper returned
a non-zero code. Basically, the macros concatentate the level with the
format string. This means that, when the level includes the ternary
operator and the conditional evaluates to the first argument, only the
level is used as the actual format string passed to printk. Only when
the conditional evaluates to the second argument, is the specified
format string actually passed through.

Fixes a regression in d184d3b32675693653d6a04f7a59b3067c725159.