drbd: fixup 'Add data-generation-UUID modification upon start of resnyc'
The problem with that commit is, that it introduced a call to
rotate_current_into_bitmap() but does not create a new current
UUID after doing the rotation.

A later call to rotate_current_into_bitmap() then does not find
a bitmap-slots that needs to be filled. If that call is from
__drbd_uuid_new_current() it then skips to create a new current
uuid!

Repair this, by considering bm_uuid slots as empty that have
the same uuid as the current primary in rotate_current_into_bitmap().

The only test of the testsuite that uncovered this bug was the
"quorum-isolate-primary" test. It did so because it provokes a
reconciliation resync and then promotes a node in a smaller
partition (new current UUID expected and necessary). Followed by
joining the partitions.
Due to the bug, this resync at joining the partitions was missing.