containers,rhel*: support for shipped modules
This adds a new mode for kernel module generation/loading that is
suitable for systems without internet access or where compiling the
module from source (in the container) is not an option such as RHCOS
(where most of the system is read only).

Most of the heavy lifting is done by a new tool[1] that find the best
matching kernel module rpm for the host kernel (uname -r).

This means we now have 4 modes:

- if a repo file is bind-mounted, packages are fetched via that repo
  (needs internet access). No host OS detection required.

- else if a node hash is specified, a repo is generated and a module
  fetched (needs internet access). If LB_DIST is set, it is used as
  distribution (within the generated repo config). If the distribution
  is not set, it needs to be detected. For that /etc/os-release has to
  be bind-mounted to /etc/host-release. If the host os is CENTOS, also
  /etc/centos-release has to be bind mounted. Avoid that type of module
  loading if possible.

- else if /usr/src is bind-mounted, the source (tarball in the container) is
  compiled from source. No host OS detection required.

- else the kos are loaded from a matching package (shipped in the
  container). No host OS detection required.

[1] https://github.com/LINBIT/python-lbdist