containers: prepare for RHEL certification
Looks like most of the work was already there, only the license was
missing. We can't just symlink the one in '..', as this would be outside
of docker's build context. I also don't want to duplicate the COPYING in
this repo. Just take one of the million copies of the GPL already in the
image. shadow-utils was chosen because it is likely that it is always
there and it specifically has a "gpl-2.0.txt", and not a "COPYING" like
other packages that in theory - if changed - would then also change our
license.