drbd: update UUIDs on target when reconciliation resync finished
When a reconciliation resync target becomes primary, it generates a new
UUID as usual. This UUID is not immediately set as the local current
UUID because of the sync target state. At the end of the resync this
UUID should be set as the local current UUID.

Promotion to primary usually does not occur during reconciliation
resyncs because the promotion logic waits for any such ongoing resyncs
to complete. However, with appropriate timing, it can occur that the
reconciliation resync starts between the check and actually becoming
primary.

It appears that this condition was only added as an optimization under
the false assumption that the current UUID would never change during a
reconciliation resync.