build: better document SPAAS situation
spatch-as-a-service is a useful tool, but it's not really documented very
clearly. Since we want to make it easy for "casual" developers to build and hack
on drbd from its git sources, we want to make sure that the requirements for
SPAAS (and building drbd in general) are clearly documented.

So, introduce a section in README.md about building drbd, which clearly outlines
what the requirements are and how to best go about building drbd.

Additionally, add some better error messages to gen_compat_patch.sh. Instead of
just failing with "no suitable spatch found", it will now display *why* it
failed (either because you're trying to build from a git checkout or because
SPAAS = false).