drbd: Fixup to recent commit "Lock UUIDs between sending and comparing..."
Commit 9f19be58a introduced a rw_sem in device. Each peer_device might have
a read lock on that rw_sem an set the HOLDING_UUID_READ_LOCK bit to
mark that.

Now, what that commit did not consider, is that before the connection
state changes to C_CONNECTED the whole connection can be deleted
via a netlink call drbd_adm_del_peer(). That in turn allows it to delete
the connection object. That frees the peer_device object. It never
goes through peer_device_disconnected(). -> it misses the up_read()
-> it locks somwhere else.

Fix that by adding the up_read() in the final free_peer_device() AND
move the call to free_peer_device() behind putting the reference
on the device, since free_peer_device() now needs the device.