drbd: Fix disconnect of diskless node with quorum enabled
At the core of this change is that when we decide to outdate the disk
state of a peer that we only do that if its current state is above
D_OUTDATED.

So far we simply also declared a D_DISKLESS node to be D_OUTDATED in
that process. Which is wrong.

At a later point the the state engine sees this state change as an
attach and produces a error code that does not realy fit to a
failure during a disconnect. Like:

c4: drbdsetup del-peer diskless-join-leave-20200310-102529 1
c4: diskless-join-leave-20200310-102529: State change failed: (-26) Intentional diskless peer may not attach a disk

The CS_VERBOSE was not missing by intention in conn_try_disconnect().
Renamed the functions to express what they do.