drbd: Fix disconnect of diskless node with quorum enabled
Recently there was already a fix in this area. The recent fix looked
at the case that the disconnect is initiated from the diskful nodes.

This fix is for the case that the diskconnect is initated from
the diskless node itself.

The error messages where:
drbdsetup down target1_lu1
State change failed: (-10) State change was refused by peer node

On the peer node it loged:
State change failed: Intentional diskless peer may not attach a disk

The reason for that is the self-outdating if a node leaves the
cluster that has a primary. But the self-outdating should not
pretend that a diskless node turns into an outdated node!

Fixing all that.