drbd: use new RB_DECLARE_CALLBACKS_MAX macro
Since the signature of RB_DECLARE_CALLBACKS changed upstream, we should
just use this new macro instead.

This is more or less the equivalent of Linux kernel upstream commit
315cc066b8ae ("augmented rbtree: add new RB_DECLARE_CALLBACKS_MAX macro").

A note on the _STATIC define: this is necessary because coccinelle
(which we use for kernel backwards compatibility) cannot deal with a
"static" in a parameter list. This could be removed if coccinelle
decides to support it.