drbd: Send config details, size, data UUIDs, and initial state before first 2pc
When establishing a connection DRBD changed connection state before
changing the states of the peer_devices. With that also the UUID comparison
took place after the first two phase commit for establishing the conneciton
(= changing connection state).
A consequence of that is that we are not able to detect a split-brain
situation when both nodes are primary. The state-engine denies the conneciton
we never come to the point where we even receive or look at the data UUIDs.

In comparison to drbd-8.4 this is a step back. 8.4 compared UUIDs first,
thus detected a split-brain situation before it even looked at the
current roles of the nodes. Users are complaining about that.

This patch changes the order of actions during establishing a connection
(only if both node are recent enough to support it).

That require to move the try_to_get_resynced() logic to the generic state
engine. To sanitize_state(), where it should be.