compat: test and patch for generic_start_io_acct
More or less a direct port from the compat.h based version.

The "too old, we don't care" case where generic_*_io_acct is not
defined, but __disk_stat_inc is, is so old that we *really* don't care
about it anymore (2.6.32 does not have it); so just drop that.