drbd: Fix a logic error in __is_valid_soft_transition()
The old implementation was wrong in that it also rejected the
connect request for connetion A if connection B has its initial
state pending. I.e. the more connections it tries to establish
in parallel the more likely it is to hit this wrong rejection
of a state change.

The intended behaviour is to only reject role changes, detach and attach
while initial states are pending.

Signed-off-by: Joel Colledge <joel.colledge@linbit.com>
Signed-off-by: Lars Ellenberg <lars.ellenberg@linbit.com>