drbd: add missing include <linux/overflow.h>
On most kernel some other include includes linux/overflow.h. On some rare
ones this does not happen (RHEL 7.6 is an example but not the only one).

Fix this by including that file.

But on older kernels that linux/overflow.h does not exist, so remove that
include by coccinelle if it remove the struct_size() macro