drbd: Allow continuation of initial resync
It was discovered that the initial resync would always re-start from the
beginning. It does not continue after an interruption.

It was like that since we got rid of the "sync-uuid" idea. That
means since drbd-9.0 (Year 2015).

In order to make it continuable, we need to replace the UUID_JUST_CREATED
with something that is in the base of bitmap of the sync source node. Since
the 2017 commit b8b46008f (released with drbd-9.0.10) such an UUID gets
generated.

With a peer that runs a DRBD before 9.0.10 it would just adopt the
UUID_JUST_CREATED value again, and the initial resync is not continuable
as it was before.

A side effect of this change is that only the resync-target sets
the bitmap and the sync-source adopts that bitmap from the sync-target.
That is okay.