drbd: remove DRBD_DEBUG_MD_SYNC
This is basically just to be more "in line" with upstream. Linux kernel
commit ac0acb9e39ac ("drbd: use drbd_device_post_work() in more places")
removed this in 2014, and I don't think anyone has really needed it since
then. So, if it's gone in upstream, remove it downstream too.

Note: this also covers the equivalent of Linux kernel commit
91dfa2dd812a ("block/drbd: delete invalid function drbd_md_mark_dirty_"),
which was basically just a fixup for the aforementioned commit.