drbd: part_stat* helpers moved from genhd.h to a new header
This is basically the equivalent of Linux kernel commit c6a564ffadc9
("block: move the part_stat* helpers from genhd.h to a new header").

Since our downstream version of drbd_int.h uses a part_stat* helper,
which is not the case in upstream, it is better for us to include it
directly through drbd_int.h instead of the individual .c files like
the original author did.