drbd: Fix a sometimes missed resync if data came in on a diskless node
This is a fixup to commit 2f144892f. We want to send one bitmap slot
that is unallocated up to now. Well, it needs to a one that does not
refer to myself and not to exactly that peer we are talking to.

The impact of the bug can be quite severe:
When this triggers e.g. a node with disk has node id 0 and a
diskless has node id 1. The current UUID of the diskful node
was set to a "day0 UUID" (by LINSTOR).
Then upon first promote on the diskless it would fail to create
a new current-uuid. Which stays undiscovered until another
diskfull node gets added (which was also set to the "day0 UUID").
At this point no resync happens!