containers: load all useful module deps
This can be considered the container that handles module loading. Mainly
DRBD, but also other modules that are useful in the context of
containerized LINSTOR.

This adds a new $LB_HOW, which can be set to "deps_only". In this mode,
it only loads the dependencies for LINSTOR layers, and exits. This might
be useful for example in a k8s operator, where the actual DRBD module is
installed on the host, but one wants to make sure all kmods are loaded
for the linstor-satellite.

We chose this container, because we already have all the required tools
for module handling installed, and the actual container and the host
have to match already. So we can be sure there are no surprises (older
versions of modprobe not being able to handle compressed modules or what
ever).