compat: Work around an spatch limitation with include files
When spatch needs to change a function declaration in a header file
it automatically also processes that header file (with all rules).
If a header file does not need a change due to no function signatures
changing no rules get applied to that header file.

We have a rule that removes #include <linux/part_stat.h> for kernels
that do not have that include files. But if there is no reason to
to touch drbd_int.h spatch misses that.