drbd: Keep the receive timeout low for the complete handshake
since we now hold the uuid_sem during a handshake it is important to
not allow that a handshake gets delayed longer than necessary.

As soon as we have a worker item that requires the uuid_sem (like
a finished resync) it will block the worker. In consequence that
blocks all state transitions.