drbd: fix a deadlock between the handshake a creating a new current-uuid
since the introduction of the uuid_sem (commit 9f19be58a) it had the
following possible deadlock:

We are in the initial handshake and have a down_read() on uuid_sem. As a part
of the handshake we receive the sizes. In receive_sizes it takes aquires
the conf_update mutex.
In case the worker threads creates a new current uuid, it takes the
conf_update mutex first and does the down_write(uuid_sem) later deep
in drbd_uuid_new_current(). -> different locking order!

Fix that by removing the requirement to hold the conf_update mutex for creating
a new current_uuid.

On May 17 2015 the conf_update was put around drbd_uuid_new_current()
only 11 days later the idea of the _ref iterators was born.

The conf_update mutex was only there to guarantee that the peer
devices stay alive while drbd_send_uuids() executes. The much
more elegant way is to hold a reference on them while
drbd_send_uuids() might sleep on a congested socket.