drbd: Obey 'allow-remote-read = no' on all new UUID decissions
The drbd_data_accessible() function considers 'allow-remote-read = no'
correctly. Remove the open coded areas and replace all of them with
calls to that function.

The recent change that made this more visible was commit
e816fc6ff7. That change introduces that upon a disk failure a
primary creates the new current uuid immediately.