drbd: Another fix for 'resync continuation'
With the last commit on that matter 42c4487c17 I introduced to set all
bitmap slots after the UUID comparison decided so.  But no! Only all
(allocated) slots if this is a fresh node! There are legit cases where
we should set only a single slot. Nicely documented in the testcase
with the name 'resync-to-older'. Thanks to that test case this
regression was found.