drbd: Fix SyncSource accepting a resync finish event
This part of the expression goes back to drbd8 times. It was changed
from D_UP_TO_DATE to D_OUTDATED in 2014 to reflect drbd9 (see ac67b5982f).
What was missed back than is that a SyncSource might even become Inconsistent.
(When it becomes SyncTarget from an other node in the mean time)

While hunting this down I created a test in the test-suite that triggered
with the wrong statement. It is called: resync-between-targets