containers: execute python setup in sub-shell
This fixes 2 "bugs". Focal has no default "python", so we should execute
python2. Second, cleanup was not executed as we did not "cd ..". Just
execute that part in a sub-shell.

For all except Focal just cleanup was missing, and for Focal all of this
isn't a real problem because we don't use the lbdist tool to find a
matching distribution, we expect a kernel module packages that perfectly
matches on Ubuntu anyways.