drbd: take uuid_sem before calling drbd_uuid_set_bitmap()
The change was originally introduced to fix a deadlock in the DRBD 8.4
compat code. DRBD 10 does not maintain compatibility with 8.4. However,
we still want this change because we want to hold the lock longer than
just in drbd_uuid_set_bitmap().

This fixes a deadlock which was introduced by skipping:
5db022ecc579 drbd: fix hardcoded deadlock in resync finished when using compat protocol
But still merging:
11aa05a2079c drbd: do not give up and immediately retake uuid_sem

---

Original commit message:

drbd: fix hardcoded deadlock in resync finished when using compat protocol

If we are SyncSource, and connected to an old DRBD (8.4),
the resync finished compat code path calls into
drbd_uuid_set_bitmap() immediately.

That tries to down_write(uuid_sem),
but since 9f19be58a drbd: Lock UUIDs between sending and comparing to the peer's version
drbd_resync_finished() has aquired it already.

-> hardcoded deadlock, always.

I try to fix this by not taking the semaphore in drbd_uuid_set_bitmap()
anymore, which means each caller has to take it or hold it already.