drbd: Fixup for conditions when a diskless primary accepts a current UUID
See also recent commit f550237a9c

With enabling it when we are primary and have the CURRENT_UUID_FLAG set,
the test_and_clear_bit() must be the last part of the expression, since
it has side effects.

But in drbd_set_exposed_data_uuid() the UUIDs are compared as well, but
there the lowest bit (UUID_PRIMARY) is stripped before the comparison.

The problem was with the lowest bit being different, we enter that
block, and clear the CURRENT_UUID_FLAG bit, but then doing nothing.
The issue is the lost CURRENT_UUID_FLAG flag bit.

Improve that by stripping the lowest bit also in the outer comparison.