drbd: Fix a corner case of a resync after a failover to a diskless node
A setup of three nodes. One diskless, two with storage. One of the
storage nodes is primary. It leaves the cluster unexpectedly. Then the
diskless node is promoted to Primary and does not touch the data-set.
(Still same current UUID).

When the node returns it establishes connections to the other two
concurrently. Let's assume it does the relevant logic with the other
storage node first. It starts a resync with the other storage node due to
having the UUID_FLAG_CRASHED_PRIMARY flag set.

It also has the same UUID as the diskless primary.

The code as it was then tried to upgrade the disk state UpToDate, but
the state sanitization removed that since it can not be UpToDate while
it is target of a resync.

The resync operation is marked as UNSTABLE at the moment it starts,
because at that time the SyncTarget did not have a connection to the
diskless primary.

The end result:
The resync finishes, the crashed and re-joined node stays Inconsistent.

This is how it looks in the log file:

drbd res peer-02: conn( Connecting -> Connected ) peer( Unknown -> Primary )
drbd res/0 drbd1005 peer-02: Upgrading local disk to D_UP_TO_DATE since current UUID matches.

And then no change to the local disk state!

The fix:
In case we find out that we have the same UUIDs as the diskless primary
but are already a resync-target, we know that we have to redo the
resync. Do that by taking the other code path.

The relevant test is called 'resync-after-failover-to-dless'.