drbd: Fix wrong hack for concurrent  reconciliation resync and promotion
Commit d9b87c3098 (from May 2019) was wrong. With that, a node changes
the peer-disk-state before it receives that update. Because it must
happen.  I was too lazy back then to do the right thing, to delay the
promotion to the point in time when it is possible.

In July 2019 with commit 72932eb98 drbd_set_role() partially got what
was deemed to be to complex back in May. Since then it delays the
state transition until the reconciliation resync is finished.

The May commit backfires at the moment the peer takes longer to
upgrade his disk state from D_CONSISTENT to D_UP_TO_DATE because it
has to retry that state change. Then the peer already allows the
promotion which gets then rejected locally.

Removing the hack from May 2019 and finish what was begun in July 2019.
Properly wait for disk to leave D_CONSISTENT state and also consider
that due to a reconciliation resync the disk state may even be
D_OUTDATED or D_INCONSISTENT.