drbd: fix connection sequence stuck in "paths crossed" state
This fixes a situation where DRBD got stuck in Connecting state,
regularly producing messages of the form:
drbd res tcp:otherhost: initial paths crossed P

The issue occurred when multiple paths to a peer are configured, and the
following occurred:
* First socket connected on path X
* No inbound connection on path X received
* Attempts to connect outbound on path X fail
* Peer is attempting to connect to us on path Y

Whenever the peer attempted to connect on path Y while we were listening
for inbound connections, the warning would be printed. The situation was
never resolved, because we were trying to establish a second socket on
path X and we closed any connections on path Y.

This occurs in the real world when someone uses the DRBD multi-path
feature as a replacement for a service IP for a stacked resource, and is
unlucky with the timing.

Fix the issue by changing how we respond when we receive a connection on
a different path to the one where we already have a socket open. Instead
of closing the new connection, we close the first socket and fail over
to the new path.