drbd: Fix abortion of local state change upon losing a connection
So far the code only handled aborting remote state changes, it failed
to properly abort it if the two-phase-commit was locally initiated.

That lead to later two-phase-commit being delayed of up to 30 seconds
(two-phase commit timeout).

There is a new test case in the test suite that highlights the issue
and was used to verify the effectiveness of this patch. It is named
'twopc_abort_connect'.