drbd: Fix a use after free that might trigger during resync
in recv_resync_read() after submitting the peer_req it may no longer
be access. It might be already destroyed!

This access after free probably did not trigger frequently. When it
triggered it caused any sync-source peers to mark an arbitrary region
of its bitmap as out-of-sync. The remaining question is if that was
the root cause of not terminating resync.

This can only be triggered if a resync-target node is (paused)
sync-source to some other node. See test case
'resync-source-becomes-target2'.