drbd: add missing locking; fix a bitmap-uuid left over
When a resync starts we rotate our current UUID into all bitmap slots
to our peers (see function drbd_uuid_resync_starting()). That is to
enable those peers to later resync from this node. That is all fine
so far.

When we successfully finish a resync (function
drbd_uuid_resync_finished()) we need to either clear or set all those
bitmap slots towards those peers.

The expression to determine which nodes' bitmap-uuid to clear
was '~ss_nz_bm | pwcu' and was later changed to '~ss_nz_bm & pwcu'.
See commits 485d3dd6 and e8529d88 .

for each bit in those 64 bit masks we have:

  ss_nz_bm   pwcu   ss_nz_bm & ~pwcu   ~ss_nz_bm | pwcu   ~ss_nz_bm & pwcu
  ---------+------+------------------+------------------+-----------------
1    0        0              0                   1                  0
2    0        1              0                   1                  1
3    1        0              1                   0                  0
4    1        1              0                   1                  0

From the table it is easy to see that with (wrong) term ~ss_nz_bm & pwcu
we do not clear all the bitmap slots we are now equal with.

In other words, if a resync finishes and we see that a peer was also
resyncing from the same sync-source (that is a bit in the ss_nr_bm)
and that peer already told us that he is on the same UUID as the one that
I am going to adopt, then we failed to clear the bitmap slot
towards that peer.

But when the also resync-target peer tells us that it finished the resync
(via receive_uuids110()) we also clear the bitmap from the sync-source
towards that peer (via update_bitmap_slot_of_peer()).

resync_finished() gets executed in the context of the worker thread on
a resync-target node. receive_uuids110() in the context of the receiver
thread.

Add missing locking.

With the locking in place pwcu = ~ss_nz_bm and both terms are now
correct. Nevertheless I am restoring the old one, which seems more
correct to me.