compat: build_cocci_cache.sh: use dirname
Use dirname instead of this shell expansion. The reason for this is that
for files in the current directory, the expansion returns the file,
while dirname returns '.'.
If we use the expansion, this means that it tries to rmdir a file over
and over again, which obviously does not work. More so, the file was
already deleted a few lines above, so it double doesn't work.