drbd: Fix one more issue of concurrent reconciliation and promotion
This is a fix related to a primary node crash. The remaining nodes
need to verify that the primary is not indirectly reachable and need
to do reconciliation resync(s). This is an addition to commit 1869f45aa3.

Consider that the events happen in this order, after the isolation of
the primary:

1) the peer notices it first  pdsk( UpToDate -> Consistent )
2) disk( UpToDate -> Consistent )
3) disk( Consistent -> UpToDate )
4) role( Secondary -> Primary )
5) should become SyncTarget of the reconciliation resync. That state
   change failes because that would make the only UpToDate disk Outdated.

Avoid all that by delaying the promotion until all of the peer disks
have left the temporary "Consistent" state.