drbd: fix fixup of resync-end
Given the following situation:
* A resync was running
* All out of sync blocks got re-synced
* The sync-target has determined that the resync is finished and sent
  the corresponding P_UUIDS110 packet
* This P_UUIDS110 packet is lost due to a network failure
Then the nodes miss the next resync, which may be necessary.

The problem is that in this case the sync-source does not clear the
bitmap-UUID towards the target. It might miss creating a new current
UUID.

In drbd-8.4 that was handled by uuid_fixup_resync_end(). For drbd9 it
is necessary to add the logic in drbd_uuid_detect_finished_resyncs()
since that cleans up the "leftover" UUIDs in the bitmap slots.